Refactor e2e test helpers into tests/helpers/ directory
Move authHelper into tests/helpers/ alongside a new organisms.ts
(split out from the old helpers.ts) so helper files are grouped
together rather than living in feature directories. Add authenticatedPage
fixture to e2e.fixture.ts so tests declare auth as a dependency instead
of calling the helper imperatively.
