+++ website/tests/e2e.fixture.ts
@@ -1,4 +1,4 @@
-import { test as base } from '@playwright/test';
+import { test as base, type Page } from '@playwright/test';
 
 import { CompareSideBySidePage } from './CompareSideBySidePage.ts';
 import { CompareToBaselinePage } from './CompareToBaselinePage.ts';
@@ -6,6 +6,7 @@ import { CompareVariantsPage } from './CompareVariantsPage.ts';
 import { SequencingEffortsPage } from './SequencingEffortsPage.ts';
 import { SingleVariantPage } from './SingleVariantPage.ts';
 import { CollectionDetailPage } from './collections/CollectionDetailPage.ts';
+import { setupAuthCookie } from './helpers/auth.ts';
 
 type E2EFixture = {
     compareVariantsPage: CompareVariantsPage;
@@ -14,6 +15,7 @@ type E2EFixture = {
     compareSideBySidePage: CompareSideBySidePage;
     singleVariantPage: SingleVariantPage;
     collectionDetailPage: CollectionDetailPage;
+    authenticatedPage: Page;
 };
 
 export const test = base.extend<E2EFixture>({
@@ -34,5 +36,9 @@ export const test = base.extend<E2EFixture>({
     },
     collectionDetailPage: async ({ page }, use) => {
         await use(new CollectionDetailPage(page));
+    },
+    authenticatedPage: async ({ page }, use) => {
+        await setupAuthCookie(page, 'e2e-test');
+        await use(page);
     },
 });
