+++ website/tests/helpers/organisms.ts
@@ -1,6 +1,6 @@
-import { allOrganisms, Organisms } from '../src/types/Organism.ts';
-import type { OrganismWithViewKey } from '../src/views/routing.ts';
-import { ServerSide } from '../src/views/serverSideRouting.ts';
+import { allOrganisms, Organisms } from '../../src/types/Organism.ts';
+import type { OrganismWithViewKey } from '../../src/views/routing.ts';
+import { ServerSide } from '../../src/views/serverSideRouting.ts';
 
 export function organismsWithView<ViewKey extends string>(viewKey: ViewKey): OrganismWithViewKey<ViewKey>[] {
     return allOrganisms.filter((organism) => ServerSide.routing.isOrganismWithViewKey(organism, viewKey));
